+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
-
+# 0.1.4 (2024-03-21)
+- Added `char` to valid types. 
 
 # 0.1.3 (2024-03-17)
-- Added Regex to valid types (no nesting, must be unique per property).
+- Added `Regex` to valid types (no nesting, must be unique per property).
 - Declares all_variants_array as static, as well as arrays of value duplicates.
 - Fixed `-o` not working, now works for dirs and files.